pyproject.toml
[tool.poetry]
name = "slavarobotam_gendiff"
version = "1.0.1"
description = "Compare two files"
authors = ["mari <zefiratakuet@yandex.ru>"]
packages = [
    { include = "gendiff" },
  ]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: Freeware"
]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.1"
pytest-cov = "^2.7"
coverage = "^4.5"
pyyaml = "^5.0"
rope = "^0.14.0"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[[tool.poetry.source]]
name = "ott45"
url = "https://test.pypi.org/legacy/"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
